requirements.txt
bcrypt==4.0.1
psycopg2==2.9.9
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dotenv==1.0.0
repackage==0.7.3
shiboken6==6.6.0